+++ config/routes.rb
@@ -1,8 +1,10 @@
 Rails.application.routes.draw do
   resources :users, only: [:new, :create]
-  get 'login', to: 'sessions#new'
-  post 'login', to: 'sessions#create'
-  get 'welcome', to: 'sessions#welcome'
-  get 'authorized', to: 'sessions#page_requires_login'
-  get 'logout', to: 'sessions#logout'
+  get '/login', to: 'sessions#new'
+  post '/login', to: 'sessions#create'
+  get '/welcome', to: 'sessions#welcome'
+  get '/authorized', to: 'sessions#page_requires_login'
+  get '/logout', to: 'sessions#logout'
+  post '/entries', to: 'entries#create'
+  get '/entries', to: 'entries#entry_buttons'
 end