+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 17/11/21
+
+* feature/add-entries
+  * User can clock in for AM Start, AM Finish, PM Start and PM Finish
+  * Four buttons used on `/entries` view for MVP
+  * Full feature tests for simulating a full day of clocking in
+
 ## 16/11/21
 
 * feature/user-authentication